+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - optim: MARS
 - optim: common file for repeated ops (e.g., newton_schulz)
+- examples: compare optimizers on mnist/cifar10 
+
+### Fixed
+- optim: ADOPT correctly implemented (exp_avg_sq and g_0)
 
 ## [v0.3.0](https://github.com/stockeh/mlx-optimizers/releases/tag/v0.3.0) - 2024-11-18